Add constants module
