+++ src/constants.ts
@@ -1,0 +1,11 @@
+export const SNACKBAR_AUTO_HIDE_DURATION = 5000;
+
+export const REPORT_ERROR_FORM_ENTRIES = {
+  url: "entry.2056392454",
+  network: "entry.1632331664",
+  platform: "entry.259085709",
+  error: "entry.1383309652",
+};
+
+export const REPORT_ERROR_FORM_URL =
+  "https://docs.google.com/forms/d/1YOvhK2RfTQmYO8DGMRqN7FYxRhBZd9jB6PZ7InJirTk/viewform";
